ease running both debug and release builds on device simultaneously
Updated build.gradle to apply package name (applicationId) suffix to
debug builds in order to differentiate it from release builds.  This
facilitates installing and running both builds on a single device at
the same time, without interfering with each other.

Also included a version name suffix to debug builds for further
clarification.

Also included resources with a debug qualifier to apply a different
app name (English and Korean provided) and launcher icon to debug
builds for further differentiation in the launcher and settings.

This should make it easier for users to debug and compare behaviors
with release builds.